linkmind_proxy/requirements.txt
flask==2.3.3
flask-cors==3.0.10
python-dotenv==1.0.0
google-generativeai==0.3.2
google-cloud-aiplatform>=1.36.0
werkzeug==2.3.7 